# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DBA Forge API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface ModuleSummary {
  id: string;
  title: string;
  description: string;
  order: number;
  isFree: boolean;
  isLocked: boolean;
  lessonCount: number;
  estimatedMinutes: number;
  topics: string[];
}

export type LessonType = (typeof LessonType)[keyof typeof LessonType];

export const LessonType = {
  scenario: "scenario",
  lesson: "lesson",
  code: "code",
  mistakes: "mistakes",
  simulation: "simulation",
  lab: "lab",
  checklist: "checklist",
} as const;

export interface Lesson {
  id: string;
  title: string;
  type: LessonType;
  content: string;
  /** @nullable */
  codeExample?: string | null;
  /** @nullable */
  options?: string[] | null;
  /** @nullable */
  correctOption?: number | null;
  /** @nullable */
  explanation?: string | null;
  /** @nullable */
  checklistItems?: string[] | null;
  order: number;
}

export interface ModuleDetail {
  id: string;
  title: string;
  description: string;
  order: number;
  isFree: boolean;
  isLocked: boolean;
  estimatedMinutes: number;
  topics: string[];
  lessons: Lesson[];
}

export interface ModuleProgress {
  moduleId: string;
  completedLessonIds: string[];
  /** @nullable */
  completedAt?: string | null;
  percentComplete: number;
}

export interface UpdateProgressBody {
  lessonId: string;
  completed: boolean;
}

export interface QuickReferenceCard {
  id: string;
  title: string;
  syntax: string;
  description: string;
  /** @nullable */
  example?: string | null;
}

export interface QuickReferenceSection {
  id: string;
  title: string;
  icon: string;
  cards: QuickReferenceCard[];
}

export interface UserProfile {
  userId: string;
  email: string;
  isPremium: boolean;
  /** @nullable */
  premiumSince?: string | null;
  modulesCompleted: number;
  totalLessonsCompleted: number;
}

export interface CheckoutSession {
  url: string;
  sessionId: string;
}

export interface PaymentVerification {
  success: boolean;
  isPremium: boolean;
  message: string;
}

export interface StatsOverview {
  totalModules: number;
  totalLessons: number;
  premiumUsers: number;
  totalEnrollments: number;
}

export type VerifyPaymentParams = {
  session_id: string;
};