# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DBA Forge API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Returns all modules with metadata and lock status for the current user
 * @summary List all training modules
 */
export const ListModulesResponseItem = zod.object({
  id: zod.string(),
  title: zod.string(),
  description: zod.string(),
  order: zod.number(),
  isFree: zod.boolean(),
  isLocked: zod.boolean(),
  lessonCount: zod.number(),
  estimatedMinutes: zod.number(),
  topics: zod.array(zod.string()),
});
export const ListModulesResponse = zod.array(ListModulesResponseItem);

/**
 * @summary Get a single module with full content
 */
export const GetModuleParams = zod.object({
  moduleId: zod.coerce.string(),
});

export const GetModuleResponse = zod.object({
  id: zod.string(),
  title: zod.string(),
  description: zod.string(),
  order: zod.number(),
  isFree: zod.boolean(),
  isLocked: zod.boolean(),
  estimatedMinutes: zod.number(),
  topics: zod.array(zod.string()),
  lessons: zod.array(
    zod.object({
      id: zod.string(),
      title: zod.string(),
      type: zod.enum([
        "scenario",
        "lesson",
        "code",
        "mistakes",
        "simulation",
        "lab",
        "checklist",
      ]),
      content: zod.string(),
      codeExample: zod.string().nullish(),
      options: zod.array(zod.string()).nullish(),
      correctOption: zod.number().nullish(),
      explanation: zod.string().nullish(),
      checklistItems: zod.array(zod.string()).nullish(),
      order: zod.number(),
    }),
  ),
});

/**
 * @summary Get all progress for the current user
 */
export const GetUserProgressResponseItem = zod.object({
  moduleId: zod.string(),
  completedLessonIds: zod.array(zod.string()),
  completedAt: zod.string().nullish(),
  percentComplete: zod.number(),
});
export const GetUserProgressResponse = zod.array(GetUserProgressResponseItem);

/**
 * @summary Update progress for a module
 */
export const UpdateProgressParams = zod.object({
  moduleId: zod.coerce.string(),
});

export const UpdateProgressBody = zod.object({
  lessonId: zod.string(),
  completed: zod.boolean(),
});

export const UpdateProgressResponse = zod.object({
  moduleId: zod.string(),
  completedLessonIds: zod.array(zod.string()),
  completedAt: zod.string().nullish(),
  percentComplete: zod.number(),
});

/**
 * Returns all quick reference sections, publicly accessible without login
 * @summary Get all DBA Quick Reference cards
 */
export const GetQuickReferenceResponseItem = zod.object({
  id: zod.string(),
  title: zod.string(),
  icon: zod.string(),
  cards: zod.array(
    zod.object({
      id: zod.string(),
      title: zod.string(),
      syntax: zod.string(),
      description: zod.string(),
      example: zod.string().nullish(),
    }),
  ),
});
export const GetQuickReferenceResponse = zod.array(
  GetQuickReferenceResponseItem,
);

/**
 * @summary Get current user profile and premium status
 */
export const GetUserProfileResponse = zod.object({
  userId: zod.string(),
  email: zod.string(),
  isPremium: zod.boolean(),
  premiumSince: zod.string().nullish(),
  modulesCompleted: zod.number(),
  totalLessonsCompleted: zod.number(),
});

/**
 * @summary Create a Stripe checkout session for full access
 */
export const CreateCheckoutSessionResponse = zod.object({
  url: zod.string(),
  sessionId: zod.string(),
});

/**
 * @summary Verify a completed payment and upgrade user to premium
 */
export const VerifyPaymentQueryParams = zod.object({
  session_id: zod.coerce.string(),
});

export const VerifyPaymentResponse = zod.object({
  success: zod.boolean(),
  isPremium: zod.boolean(),
  message: zod.string(),
});

/**
 * Returns overall stats like total users, completions, module count
 * @summary Get platform stats overview
 */
export const GetStatsOverviewResponse = zod.object({
  totalModules: zod.number(),
  totalLessons: zod.number(),
  premiumUsers: zod.number(),
  totalEnrollments: zod.number(),
});