# lib/api-zod/src/generated/types/checkoutSession.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DBA Forge API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CheckoutSession {
  url: string;
  sessionId: string;
}