# lib/api-zod/src/generated/types/lesson.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DBA Forge API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LessonType } from "./lessonType";

export interface Lesson {
  id: string;
  title: string;
  type: LessonType;
  content: string;
  /** @nullable */
  codeExample?: string | null;
  /** @nullable */
  options?: string[] | null;
  /** @nullable */
  correctOption?: number | null;
  /** @nullable */
  explanation?: string | null;
  /** @nullable */
  checklistItems?: string[] | null;
  order: number;
}