# lib/api-zod/src/generated/types/moduleDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DBA Forge API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Lesson } from "./lesson";

export interface ModuleDetail {
  id: string;
  title: string;
  description: string;
  order: number;
  isFree: boolean;
  isLocked: boolean;
  estimatedMinutes: number;
  topics: string[];
  lessons: Lesson[];
}