# lib/api-zod/src/generated/types/paymentVerification.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DBA Forge API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PaymentVerification {
  success: boolean;
  isPremium: boolean;
  message: string;
}