# lib/api-zod/src/generated/types/quickReferenceSection.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DBA Forge API specification
 * OpenAPI spec version: 0.1.0
 */
import type { QuickReferenceCard } from "./quickReferenceCard";

export interface QuickReferenceSection {
  id: string;
  title: string;
  icon: string;
  cards: QuickReferenceCard[];
}