# lib/api-zod/src/generated/types/userProfile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DBA Forge API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UserProfile {
  userId: string;
  email: string;
  isPremium: boolean;
  /** @nullable */
  premiumSince?: string | null;
  modulesCompleted: number;
  totalLessonsCompleted: number;
}